tools/perf/scripts: Fix pylint warnings on mem_phys_addr.py
Patch fixes pylint warnings bad indentation, unused import,
snake_case naming convention, missing doc_string, string formating.
